Refactor MainHeader component structure and styles; separate background into its own component and CSS module
